Made layering more sane, more client interactions, updates to interactable UI and UIFocusable interface
Big thing in this commit is the changes to layering (UIComponents can
now, rightfully, handle events instead of just receiving them), and the
UIFocusable interface. The UIFocusable interface is meant to be used by
UI elements as a way to determine where KeyEvents need to go when they
are dispatched. Specifically, one element at a time gets the focus, and
they are the only UI element in a 'focused' (or maybe just top layer)
panel that the events are sent to. This is a way for me to bring in my
custom textfield, and later to expand on buttons. For examble, I might
implement using tab to switch between focusable elements.